continent map legend and labels
