+++ WORK_SUMMARY.md
@@ -46,3 +46,17 @@ Time: ~5 hours
 -   Add basic styling to the app using reference from cathode-ray-tube monitors and vintage air traffic control systems. (see reference images in [./research/images/maps](./research/images/maps) and colour theme generation in [./research/images/maps/colourscheme.jpg](./research/images/maps/colourscheme.jpg)) ![Colour scheme extraction from reference](./research/images/maps/colourscheme.jpg). **Note: this is a departure from the original style research done in [Section 2: Brainstorming](#2-brainstorming), but is a better fit for the vector-based display of the map data (and quicker to implement and get looking good).** ![First pass at styling](./research/styled-world-map.png)
 -   Form elements for editing trip withing the trip overview component.
 -   Change start-city-selection to be triggered by clicking anywhere on a country. Clicks can be less precise in general, but allows for greater precision when selecting cities that are close together (previous option relied on expanded "hit box" for each city, that *could* overlap and block each other).
+
+---
+
+## Things to improve if I had more time
+
+* Zoom in on trips (currently the map is always shows a full continent)
+* Resolve geo-data source (there are geo-political aspects I'm not qualified to judge)
+* Incorporate more options for route calculation: e.g. "same currency", "same languages spoken" etc.
+* Either combine the geoJSON data with the API data or access the geoJSON data from an API rather than yoloing it with the [/src/js/data](./src/js/data) files.
+* Consolidate the two map components into a single component - there's a lot of duplication between the two.
+* Setup better state-management for trip data (most likely using a flux model with `useReducer` and `useContext` hooks).
+* Better keyboard support for selecting items on the map (switch from tabbing to using the arrow keys, for instance).
+* Resolve overlapping city markers for routes that are close together (the Caribbean is a good example of this - currently an illegible mess).
+* Small screen layout (currently *no* thought has been given to mobile layout).
