+++ src/scss/app.scss
@@ -22,6 +22,7 @@
 @import "component.content";
 @import "component.header";
 @import "component.map";
+@import "component.trip-overview";
 
 // TRUMPS
 @import "trumps.layout";
